+++ src/libtiled/object.h
@@ -167,15 +167,20 @@ public:
     static void setPropertyTypes(const SharedPropertyTypes &propertyTypes);
     static const PropertyTypes &propertyTypes();
 
+    QObject *editable() const { return mEditable; }
+
+protected:
+    /**
+     * The editable wrapper created for this object. Deleted along with the
+     * object, unless it is owned by the script engine.
+     */
+    QPointer<QObject> mEditable;
+
 private:
     const TypeId mTypeId;
     QString mClassName;
     Properties mProperties;
 
-    /**
-     * The editable wrapper created for this object.
-     */
-    QPointer<QObject> mEditable;
     friend class EditableObject;
 
     static SharedPropertyTypes mPropertyTypes;