+++ src/tiled/editabletileset.h
@@ -112,6 +112,7 @@ public:
 
     Q_INVOKABLE explicit EditableTileset(const QString &name = QString(),
                                          QObject *parent = nullptr);
+    explicit EditableTileset(const SharedTileset &tileset, QObject *parent = nullptr);
     explicit EditableTileset(const Tileset *tileset, QObject *parent = nullptr);
     explicit EditableTileset(TilesetDocument *tilesetDocument,
                              QObject *parent = nullptr);
@@ -206,6 +207,8 @@ private:
 
     void wangSetAdded(Tileset *tileset, int index);
     void wangSetRemoved(WangSet *wangSet);
+
+    friend class TilesetDocument;
 
     bool mReadOnly = false;
     SharedTileset mTileset;